feat: clean wb-updates & centralize config

+++ src/seismogram_pipeline/cli/resize_image.py
@@ -18,25 +18,29 @@ from docopt import docopt
 from skimage.transform import resize
 import imageio
 
+
 def resize_image(in_file, scale, out_file):
-  from ..core.load_image import get_image
-  from ..core.timer import timeStart, timeEnd
+    from ..core.load_image import get_image
+    from ..core.timer import timeStart, timeEnd
 
-  timeStart("load image")
-  image = get_image(in_file)
-  timeEnd("load image")
+    timeStart("load image")
+    image = get_image(in_file)
+    timeEnd("load image")
 
-  timeStart("resize image")
-  new_shape = (int(image.shape[0] * scale), int(image.shape[1] * scale))
-  resized = resize(image, new_shape, preserve_range=True, anti_aliasing=True).astype(image.dtype)
-  timeEnd("resize image")
+    timeStart("resize image")
+    new_shape = (int(image.shape[0] * scale), int(image.shape[1] * scale))
+    resized = resize(image, new_shape, preserve_range=True, anti_aliasing=True).astype(
+        image.dtype
+    )
+    timeEnd("resize image")
 
-  if out_file:
-    timeStart("save image")
-    imageio.imwrite(out_file, resized)
-    timeEnd("save image")
-  else:
-    print("No output file specified.")
+    if out_file:
+        timeStart("save image")
+        imageio.imwrite(out_file, resized)
+        timeEnd("save image")
+    else:
+        print("No output file specified.")
+
 
 def main():
     """Main entry point for the resize_image CLI."""
@@ -50,5 +54,6 @@ def main():
     else:
         print(arguments)
 
-if __name__ == '__main__':
+
+if __name__ == "__main__":
     main()
